+++ src/main/java/com/backend/allreva/module/concert/concert/application/query/port/ConcertSearchRepository.java
@@ -1,4 +1,4 @@
-package com.backend.allreva.module.concert.concert.application.port;
+package com.backend.allreva.module.concert.concert.application.query.port;
 
 import com.backend.allreva.common.pagination.SliceResponse;
 import com.backend.allreva.module.concert.concert.application.dto.ConcertThumbnail;
