+++ src/main/java/com/backend/allreva/module/recruitment/rent/application/query/port/RentSearchRepository.java
@@ -1,4 +1,4 @@
-package com.backend.allreva.module.recruitment.rent.application.port;
+package com.backend.allreva.module.recruitment.rent.application.query.port;
 
 import com.backend.allreva.common.pagination.SliceResponse;
 import com.backend.allreva.module.recruitment.rent.application.query.dto.RentThumbnail;
