+++ src/main/java/com/backend/allreva/module/recruitment/survey/application/query/port/SurveySearchRepository.java
@@ -1,4 +1,4 @@
-package com.backend.allreva.module.recruitment.survey.application.port;
+package com.backend.allreva.module.recruitment.survey.application.query.port;
 
 import com.backend.allreva.common.pagination.SliceResponse;
 import com.backend.allreva.module.recruitment.survey.application.dto.SurveyThumbnail;
